requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==25.3.0
    # via aiohttp
certifi==2025.7.9
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
colorama==0.4.6
    # via tqdm
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via anyio
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
httpx-sse==0.4.1
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.26
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.26
    # via -r requirements.in
langchain-core==0.3.66
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.27
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via
    #   -r requirements.in
    #   langchain
langsmith==0.4.5
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.6.1
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==1.26.4
    # via langchain-community
openai==1.95.1
    # via langchain-openai
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via langchain-community
python-dotenv==1.1.1
    # via pydantic-settings
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.11.6
    # via tiktoken
requests==2.32.4
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   langchain
    #   langchain-community
tenacity==8.5.0
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.9.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   anyio
    #   exceptiongroup
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
yarl==1.20.1
    # via aiohttp
zstandard==0.23.0
    # via langsmith
